Fixed bug due to lua closure semantics
Because startLaser() was not a part of self, the CreateThread call was globalizing the scope (or at least pulling from the global namespace), leading to multiple lasers using the same values if setActive(true) was called after creating many in a row.

I originally kept startLaser() as not a part of self to privatize it because it's an implementation function, but to do that and fix the closure issue, you'd have to pass every one of the objects variables to the function, and that is a bit ridiculous. So I added it to self and prepended an underscore to the name to make a bit more clear that it's not part of the public API of the object.