Fix infinite loop when laser has one target point
The logic for moving the laser between target points would either cause a resource crash (if randomTargetSelection was false) or an infinite loop and a game crash (if randomTargetSelection was true). To fix this bug and to improve performance in the case of one target point, a special case was added to _startLaser() that does no movement code and instead of draws the laser if there is only one target point.