Add collision caching
Collision caching works by having a set collisionGranularity, which is basically how many collisions you cache per transition from point to point. So if it is 0.01, then you cache at most 100 collisions. When the interpolated point is being calculated, you take the percentOfTravelTime and find the closest value to it based on the granularity to get a collisionIndex. You then check in the collisionCache for that particular point transition (fromIndex, toIndex, collisionIndex), and if it has a value, you return it, otherwise you do the RayCast, get the value, cache and return it.

Overtime, as more unique point transitions occur, the cache will fill up. The more full the cache is, the lower the cost because the less raycasting you'll be doing.