Add laser collision with player
This would be widly too expensive if done naively, because of the cost of ray casting. Instead, we use math to determine how close the player is to the line that represents the laser, and if it is under a threshold value (currently 0.5m), then we do a ray cast to determine if the player is actually being hit by the laser.