Add extensionEnabled option
This option allows you to disable the default behavior where we extend the laser a long distance (even through geometry) to try and prevent mid-air stops of the laser because of the linear interpolation. This is probably safe to set to false if you target points are only ever on the floor.